refactor(frontend): decouple business logic from ProductivityScreen
Debt Identified: The `ProductivityScreen` in `frontend/app/(main)/productivity.tsx` was a "God Class" exceeding 850 lines, heavily mixing UI rendering with complex data filtering, sorting, and derived state calculation logic for notes, tasks, and events.

Refactored Code: Extracted the data manipulation logic into a dedicated custom hook `useProductivityData` at `frontend/src/hooks/useProductivityData.ts`.

Structural Note: This adheres to the SOLID Single Responsibility Principle by isolating business logic from the UI layer, improving code readability, maintainability, and testability without altering existing functionality.
